Demo/requirements.txt
# Quantum - 量子化学数据处理工具依赖
# 生成日期: 2026-03-08

# ==================== GUI 框架 ====================
PyQt5>=5.15.0

# ==================== 数据处理 ====================
pandas>=1.3.0
openpyxl>=3.0.0          # Excel xlsx 读写支持

# ==================== 文档处理 ====================
python-docx>=0.8.11      # Word 文档生成

# ==================== 可选依赖 ====================
# pyinstaller>=5.0.0     # 打包为 exe (如需打包可取消注释)
# numpy>=1.20.0          # 数值计算 (部分测试文件使用)
